pysims/requirements.txt
airspeed==0.5.4.dev20150515
cachetools==0.8.0
calmap==0.0.6
cdo==1.3.0
certifi==0.0.8
chardet==1.0.1
click==6.6
click-plugins==1.0.3
cligj==0.4.0
colorama==0.3.6
cycler==0.10.0
Fiona==1.6.3.post1
geojson==1.3.2
matplotlib==1.5.1
nco==0.0.2
netCDF4==1.2.2
numpy==1.10.4
pandas==0.17.1
pyparsing==2.1.1
python-dateutil==2.4.2
pytz==2015.7
PyYAML==3.11
requests==2.9.1
Rtree==0.8.2
ruamel.base==1.0.0
ruamel.ordereddict==0.4.9
ruamel.yaml==0.11.1
scipy==0.17.0
Shapely==1.5.15
six==1.10.0
wget==3.2
xarray==0.7.1
XlsxWriter==0.8.4
